+++ app/domain/exceptions.py
@@ -73,14 +73,3 @@ class SearchDataRetrievalError(OverfastError):
     """Generic search data retrieval Exception (namecards, titles, etc.)"""
 
     message = "Error while retrieving search data"
-
-
-class BlizzardTimeoutError(Exception):
-    """Raised by the worker when a Blizzard request times out (HTTP 504).
-
-    Used as a precise filter for the taskiq retry middleware so that only
-    genuine Blizzard timeouts are retried, not other HTTP errors (404, 503…).
-    """
-
-    def __str__(self) -> str:
-        return "Blizzard request timed out"
